backend/requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.32,<1.0
sqlalchemy>=2.0,<3.0
pydantic-settings>=2.6,<3.0
pillow>=11.0,<12.0
exifread>=3.5,<4.0
psycopg[binary]>=3.2,<4.0
rawpy>=0.25,<1.0
opencv-python-headless>=4.11,<5.0
